Add parser from CSV to object structure (#60)
We have all experienced over the last year how difficult it can be to get the right data out of the Dataset that is read from the instructions.csv file. Sometimes it is in a column based view, sometimes it is row based. Also the amount of checks that were done in the preconditions of the AddDatasetMetadataToDeposit action was becoming too large and very much cluttered. It raised many issues over time like ' is mandatory and should be tested as such' or 'there is only one value per dataset allowed for '. With the setup that we had of just maintaining a Dataset as a Map[String, List[String]] it meant often these things were tested multiple times, once in the preconditions and once in the action itself.

Luckily we have Scala's type system to the rescue! You can declare a lot of things in the types of a class already. Just two examples:

a field is optional - make it an Option
a list has to have at least one element - make it a NonEmptyList
With these things I set out to rewrite the whole CSV header sequence with all its implicit rules into an object structure that is typesafe and therefore guarantees us so many things that we weren't able to be sure of before.
The new CSV reader parses the instructions.csv into an object structure that is immediately validated with all the rules (and more) that were already in AddDatasetMetadataToDeposit. It returns a Dataset (object/case class), which can be passed to the various actions that are defined in this module. These can now get the necessary data from this object.

Example: getting the DDM_CREATED date in one of the actions now doesn't require to (1) look up this column, (2) check there is only one NON-BLANK value in it, (3) check that this value actually represents a VALID date, (4) ... (5) ..., but simply call dataset.profile.created.

* changed allfields example

* delete old test datasets

* changed failing tests

* small changes to instructions.csv

* remove old headers from tests

* give the correct path to the subtitles file in files.xml

* add multiple subtitles for video to the example

* adjust test for changed example output

* accessibleTo accessrights

* upgrade ddm scheme from 2012/11 to 2016

* forgotten schema change in the test

* remove old test multi-deposits from debug-reset-apphome script

* additions and checks of AV_* and SF_* columns

* headers

* headers license

* output and input directory confusion

* another change to the instructions.csv

* SF_USER isn't optional anymore

* SF_ACCESSIBILITY and DDM_ACCESSRIGHT checks

* license header

* correct mistakes in output path of files in ruimtereis02

* add springfield fields to properties

* fix dataset 3

* only compute mimetypes if directory exists

* make sure the springfield fields are not in the properties if the dataset is not A/V

* solve todos about monad-transformers

* documentatie

* ddm version fix

* validHeaders is not a list of datasetIds; thanks @paulboon!

* refactor nested maps

* mimetype type alias suggested by @paulboon

* parser test

* put available with default in correct place in object model

* add A/V specific metadata to files.xml

* fix and expand tests for AddFileMetadataToDeposit

* delete Springfield code

* extra tests for STREAMING_SURROGATE_RELATION

* remove unnecessary springfield references from the code

* review suggestions by @lindareijnhoudt

* review suggestions by @jo-pol

* rename maybe to opt

* profile parsing completed

* parse metadata

* fix bugs in former commit

* finished basic parsing

* extract new parser

* format and rearrange

* WIP; the tests fail

* fixed and rewritten all tests

* fix todos

* test combine and replace DatasetID with DatasetId

* final cleanup

* update example

* temporary fix of a virus scanner problem

* fix parser error output in the console

* extra file exists check

* add logging statement for csv parser

* split ActionException and ParseException

* license

* ldap documentation fix